# Add UL logging of deoptimization event