--- a/src/hotspot/share/runtime/deoptimization.cpp
+++ b/src/hotspot/share/runtime/deoptimization.cpp
@@ -1838,19 +1838,20 @@ static void post_deoptimization_event(CompiledMethod* nm,
 
 #endif // INCLUDE_JFR
 
-void Deoptimization::print_ul(CompiledMethod* nm, intptr_t pc, frame& fr, int trap_bci, int osr_bci,
+void Deoptimization::print_ul(CompiledMethod* nm, intptr_t pc, frame& fr, int trap_bci,
                               const char* reason_name, const char* reason_action) {
   LogTarget(Debug, deoptimization) lt;
   bool is_osr = nm->is_osr_method();
-  if(lt.is_enabled()) {
+  if (lt.is_enabled()) {
     LogStream ls(lt);
     ls.print("cid=%d %s pc=" INTPTR_FORMAT " relative_pc=" INTPTR_FORMAT " level=%d",
-             nm->compile_id(), (is_osr ? "%" : ""), pc, fr.pc() - nm->code_begin(), nm->comp_level());
+             nm->compile_id(), (is_osr ? "%" : ""), pc, fr.pc() - nm->code_begin(),
+             nm->comp_level());
     nm->method()->print_short_name(&ls);
     ls.print(" @ ");
     ls.print("%d ", trap_bci);
-    if(is_osr) {
-      ls.print("%d ", osr_bci);
+    if (is_osr) {
+      ls.print("%d ", nm->osr_entry_bci());
     }
     ls.print("%s ", reason_name);
     ls.print("%s", reason_action);
@@ -1952,19 +1953,16 @@ JRT_ENTRY(void, Deoptimization::uncommon_trap_inner(JavaThread* current, jint tr
       get_method_data(current, profiled_method, create_if_missing);
 
     { // Log Deoptimization event for JFR, UL and event system
-      bool is_osr = nm->is_osr_method();
-      int osr_bci = is_osr ? nm->osr_entry_bci() : -1;
       Method* tm = trap_method();
-      char* name_sig = tm->name_and_sig_as_C_string();
       const char* reason_name = trap_reason_name(reason);
       const char* reason_action = trap_action_name(action);
       intptr_t pc = p2i(fr.pc());
 
       JFR_ONLY(post_deoptimization_event(nm, tm, trap_bci, trap_bc, reason, action);)
-      print_ul(nm, pc, fr, trap_bci, osr_bci, reason_name, reason_action);
+      print_ul(nm, pc, fr, trap_bci, reason_name, reason_action);
       Events::log_deopt_message(current, "Uncommon trap: reason=%s action=%s pc=" INTPTR_FORMAT " method=%s @ %d %s",
                                 reason_name, reason_action, pc,
-                                name_sig, trap_bci, nm->compiler_name());
+                                tm->name_and_sig_as_C_string(), trap_bci, nm->compiler_name());
     }
 
     // Print a bunch of diagnostics, if requested.