--- a/src/hotspot/share/runtime/deoptimization.hpp
+++ b/src/hotspot/share/runtime/deoptimization.hpp
@@ -165,12 +165,11 @@ public:
 #endif
 
   private:
+    static void print_ul(CompiledMethod* nm, intptr_t pc, frame& fr, int trap_bci,
+                         const char* reason_name, const char* reason_action);
 
-  static void print_ul(CompiledMethod* nm, intptr_t pc, frame& fr, int trap_bci, int osr_bci,
-                       const char* reason_name, const char* reason_action);
-
-  // Does the actual work for deoptimizing a single frame
-  static void deoptimize_single_frame(JavaThread* thread, frame fr, DeoptReason reason);
+    // Does the actual work for deoptimizing a single frame
+    static void deoptimize_single_frame(JavaThread* thread, frame fr, DeoptReason reason);
 
 #if COMPILER2_OR_JVMCI
   // Deoptimize objects, that is reallocate and relock them, just before they