This PR adds a new "deoptimization" tag and a new log message on the info level mimicking JFR's deoptimization event. In the future this tag will be used to add trace level information, hopefully replacing TraceDeoptimization. JFR's event also captures the stacktrace, which this does not (happy to add this if considered important).

An example output looks like this: 

```
[1.346s][info][deoptimization] Deoptimization start {
[1.346s][info][deoptimization]   compileId = 407
[1.346s][info][deoptimization]   compiler = c2
[1.346s][info][deoptimization]   method:lineno:bci = java.util.concurrent.locks.ReentrantLock$Sync.tryRelease(I)Z:176:26
[1.346s][info][deoptimization]   instruction = ifne
[1.346s][info][deoptimization]   reason = unstable_if
[1.346s][info][deoptimization]   action = reinterpret
[1.346s][info][deoptimization] } Deoptimization end
```