@JornVernee 2022-05-20
Thanks for looking at this!

I suggest trying to match the style of the `-Xlog:jit+compilation=debug` messages, since deoptimization is like the dual of compilation. And trying to keep the output on a single line,

Rather than printing the compiler name, I think printing the compilation level is more useful, as well as including some info about whether it was an osr compilation maybe. The `jit+compilation` log messages do the latter by printing an extra `%`.

Maybe something like:

```
<compile id> <is osr> <compile level> <method signature> @ <bci> <reason> -> <action>
```

The instruction and line number could also be incorporated somewhere, but the compilation messages don't have them either. They seem fine to leave out I think.

Also, I noticed that `jit+compilation` has the `debug` level, so maybe this log message should be under the `debug` level as well, instead of `info`. I don't know if there is any guidance on that somewhere.