@JornVernee 2022-05-24
> I'm not sure whether we can know if something was an OCR compilation or not.

`CompiledMethod` has the function `is_osr_method` which I think can be used for that. The code in question seems to be dealing with a `CompilerMethod*` already (https://github.com/openjdk/jdk/blob/master/src/hotspot/share/runtime/deoptimization.cpp#L1876)