@jdksjolen 2022-05-27
I changed the log to one line and moved out all repeatedly used values into variables. I limited all of these to a scope to not pollute the function scope unnecessarily. I use `comp_level` directly instead of translating into a string (didn't find a function for it), are levels as ints sufficient?