@dholmes-ora 2022-06-02
My comments were only "drive-by" observations. You really need the compiler folk to perform the actual review here.