@vnkozlov 2022-06-09
I noticed that method name is missing signature (arguments type) which we have in `TraceDeoptimization` output:
```
UNCOMMON TRAP method=java.util.Properties.getProperty(Ljava/lang/String;)Ljava/lang/String;  bci=10 pc=0x00007fbae0c23a94, relative_pc=0x00000000000000b4, debug_id=0 compiler=c2 compile_id=9 (@0x00007fbae0c23a94) thread=2203929 reason=null_check action=make_not_entrant unloaded_class_index=-1 debug_id=0
```
Yes, `PrintCompilation` does not have it but it would be nice to have it in this output.