@vnkozlov 2022-06-09
Please, also update the output example in PR's description. It is misleading now.