@navyxliu 2022-06-09
hi, @jdksjolen , 
Could you update PR description example?  It looks like you have changed it to a one-liner.  just for a record in case we need to revisit this PR.  

Do you consider 'deopt' instead of 'deoptimization'? I feel deoptimization is too long as a tag.  Neither of them is a word anyway. :)