@jdksjolen 2022-06-10
I reverted to using name_and_sig for output, updated the PR and fixed the nits.

@navyxliu, I did consider moving to `deopt` but opted against it as there are just as long tags in the rest of the set (for example "defaultmethods" and "methodcomparator").