@vnkozlov 2022-06-10
> @navyxliu, I did consider moving to `deopt` but opted against it as there are just as long tags in the rest of the set (for example "defaultmethods" and "methodcomparator").

Technically speaking it should be `uncommon_trap_deopt` because it does not cover other case of deoptimization when compiled method deoptimized due to a class change (new related class loading, unloaded, redefined) or re-compilation (from C1 to C2). That path [fetch_unroll_info()](https://github.com/openjdk/jdk/blob/master/src/hotspot/share/runtime/deoptimization.cpp#L163) is not covered by this change (may be another RFE).